lora/requirements.txt
adafruit-circuitpython-rfm69
balena-sdk
Flask
gps3
Paste
pycrypto
RPi.GPIO
spidev
toml
waitress
